llm-agent/requirements.txt
rich
wikipedia
langchain
langchain-community
langchain-openai
python-dotenv
duckduckgo-search
langchain_experimental
langchain_litellm
smolagents
smolagents[telemetry]
arize-phoenix-otel
openinference-instrumentation-smolagents